financial analyst/requirements.txt
crewai>=0.12.2
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
yfinance>=0.2.35
pandas>=2.1.0
plotly>=5.18.0 